# core-service/api/openapi/v1.yaml
openapi: 3.0.0
info:
  title: CRM Core Service API
  description: API for managing CRM core objects (Accounts, Contacts, Opportunities, Notes)
  version: 1.0.0
servers:
  - url: /v1/api
    description: Base API path
paths:
...
  /opportunities:
    get:
      summary: List all opportunities
      operationId: listOpportunities
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of opportunities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Opportunity'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create a new opportunity
      operationId: createOpportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityInput'
      responses:
        '201':
          description: Opportunity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /opportunities/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get opportunity by ID
      operationId: getOpportunity
      responses:
        '200':
          description: Opportunity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      summary: Update an opportunity
      operationId: updateOpportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityInput'
      responses:
        '200':
          description: Opportunity updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete an opportunity
      operationId: deleteOpportunity
      responses:
        '204':
          description: Opportunity deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /opportunities/account/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get opportunities by account ID
      operationId: getOpportunitiesByAccount
      responses:
        '200':
          description: A list of opportunities for the account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Opportunity'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # Note Endpoints
  /notes:
    get:
      summary: List all notes
      operationId: listNotes
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of notes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create a new note
      operationId: createNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /notes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get note by ID
      operationId: getNote
      responses:
        '200':
          description: Note details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      summary: Update a note
      operationId: updateNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '200':
          description: Note updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a note
      operationId: deleteNote
      responses:
        '204':
          description: Note deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /notes/record/{type}/{id}:
    parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum: [account, contact, opportunity]
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get notes by record ID and type
      operationId: getNotesByRecord
      responses:
        '200':
          description: A list of notes for the record
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /notes/associations:
    post:
      summary: Create note association
      operationId: createNoteAssociation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteAssociation'
      responses:
        '201':
          description: Note association created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteAssociation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete note association
      operationId: deleteNoteAssociation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteAssociation'
      responses:
        '204':
          description: Note association deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  schemas:
...
    Opportunity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        opportunity_name:
          type: string
        account_id:
          type: string
          format: uuid
        primary_contact_id:
          type: string
          format: uuid
        stage:
          type: string
          enum: [Prospecting, Qualification, Negotiation, Closed Won, Closed Lost]
        amount:
          type: number
          format: double
        close_date:
          type: string
          format: date
        probability:
          type: number
          format: double
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - opportunity_name
        - account_id
        - stage
    OpportunityInput:
      type: object
      properties:
        opportunity_name:
          type: string
        account_id:
          type: string
          format: uuid
        primary_contact_id:
          type: string
          format: uuid
        stage:
          type: string
          enum: [Prospecting, Qualification, Negotiation, Closed Won, Closed Lost]
        amount:
          type: number
          format: double
        close_date:
          type: string
          format: date
        probability:
          type: number
          format: double
          minimum: 0
          maximum: 100
      required:
        - opportunity_name
        - account_id
        - stage
    Note:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        associations:
          type: array
          items:
            $ref: '#/components/schemas/NoteAssociation'
      required:
        - id
        - content
    NoteInput:
      type: object
      properties:
        content:
          type: string
        created_by:
          type: string
          format: uuid
        associations:
          type: array
          items:
            $ref: '#/components/schemas/NoteAssociation'
      required:
        - content
    NoteAssociation:
      type: object
      properties:
        note_id:
          type: string
          format: uuid
        record_id:
          type: string
          format: uuid
        record_type:
          type: string
          enum: [account, contact, opportunity]
      required:
        - note_id
        - record_id
        - record_type
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error' 